third_party/blink/web_tests/platform/generic/fast/events/selectstart-by-drag-expected.txt

This test ensures selectstart is fired exactly once when selecting text by a mouse drag.

Initial state: PASS
Mouse down: PASS
Moving slightly to the right: PASS
Moving slightly to the left: PASS
Moving to the right: PASS
Moving further to the right: PASS
Moving back to the left: PASS
Moving to the right again: PASS
Mouse down on the right: PASS
Moving to the left: PASS
Done.
